feat: Add Footer component and related types and utilities for enhanced dashboard navigation
